pyproject.toml
[project]
name = "networking-protocol-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "paho-mqtt>=2.1.0",
    "websocket>=0.2.1",
    "websockets>=15.0.1",
]
